requirements.txt
# Django and REST framework
django>=5.1.7
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-safedelete>=1.3.2
django-cors-headers>=4.3.1

# Database
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9

# Data processing
polars>=0.20.2

# Storage
b2sdk>=1.24.0
boto3>=1.34.14

# Environment and configuration
python-dotenv>=1.0.1

# Asynchronous tasks
celery>=5.3.6
redis>=5.0.1

# Web server
gunicorn>=21.2.0
whitenoise>=6.6.0

# Utilities
requests>=2.31.0
pyjwt>=2.9.0
